Node.Update(int) -> Node.Update(float)
git-svn-id: svn://srv-dc/svn2/Citrus@190682 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/Application/Application.cs
@@ -251,7 +251,7 @@ namespace Lime
 
 		public virtual void OnCreate() {}
 		public virtual void OnTerminate() {}
-		public virtual void OnUpdateFrame(int delta) {}
+		public virtual void OnUpdateFrame(float delta) {}
 		public virtual void OnRenderFrame() {}
 		public virtual void OnDeviceRotating(DeviceOrientation toOrientation) {}
 		public virtual void OnDeviceRotated(DeviceOrientation fromOrientation) {}
